+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## [Unreleased]
+## [0.2.0] - 2019-06-01
+### Changed
+- Connection use hash-style args
 ## [0.1.2] - 2019-05-13
 ### Added
 - delegators in transaction and pledge